@@ -4,7 +4,7 @@ description: A new Peercoin wallet.
 version: 1.2.3+135
 
 environment:
-  sdk: '>=3.0.0 <4.0.0'
+  sdk: '>=3.2.0 <4.0.0'
   
 publish_to: none
 
@@ -68,7 +68,7 @@ dev_dependencies:
   build_runner: ^2.0.4
   flutter_launcher_icons: ^0.11.0
   flutter_native_splash: ^2.2.7
-  flutter_lints: ^2.0.1
+  flutter_lints: ^3.0.0
 
 flutter_icons:
   android: "launcher_icon"
