@@ -9,12 +9,12 @@ class PeerButton extends StatelessWidget {
   final bool small;
   final bool active;
   const PeerButton({
-    Key? key,
+    super.key,
     required this.text,
     required this.action,
     this.small = false,
     this.active = true,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -49,8 +49,7 @@ class PeerButton extends StatelessWidget {
 class PeerButtonBorder extends StatelessWidget {
   final Function() action;
   final String text;
-  const PeerButtonBorder({Key? key, required this.text, required this.action})
-      : super(key: key);
+  const PeerButtonBorder({super.key, required this.text, required this.action});
 
   @override
   Widget build(BuildContext context) {
@@ -92,11 +91,11 @@ class PeerButtonSetup extends StatelessWidget {
   final bool active;
 
   const PeerButtonSetup({
-    Key? key,
+    super.key,
     required this.text,
     required this.action,
     this.active = true,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -138,11 +137,11 @@ class PeerButtonSetupLoading extends StatelessWidget {
   final bool loading;
 
   const PeerButtonSetupLoading({
-    Key? key,
+    super.key,
     required this.text,
     required this.action,
     this.loading = false,
-  }) : super(key: key);
+  });
 
   @override
   Widget build(BuildContext context) {
@@ -193,7 +192,7 @@ class PeerButtonSetupLoading extends StatelessWidget {
 }
 
 class PeerButtonSetupBack extends StatelessWidget {
-  const PeerButtonSetupBack({Key? key}) : super(key: key);
+  const PeerButtonSetupBack({super.key});
 
   @override
   Widget build(BuildContext context) {
